Moved commands into separate classes
